[Project64] Make sure the x86 recompiler code is used just on x86

+++ Source/Project64-core/N64System/Recompiler/CodeBlock.h
@@ -16,7 +16,7 @@
 class CCodeBlock
 {
 public:
-    CCodeBlock(uint32_t VAddrEnter, uint8_t * CompiledLocation );
+    CCodeBlock(uint32_t VAddrEnter, uint8_t * CompiledLocation);
     ~CCodeBlock();
 
     bool Compile();
@@ -25,7 +25,7 @@ public:
     uint32_t    VAddrFirst() const { return m_VAddrFirst; }
     uint32_t    VAddrLast()  const { return m_VAddrLast; }
     uint8_t *   CompiledLocation() const { return m_CompiledLocation; }
-    int32_t     NoOfSections() const { return m_Sections.size() - 1;}
+    int32_t     NoOfSections() const { return (int32_t)m_Sections.size() - 1; }
     const CCodeSection & EnterSection() const { return *m_EnterSection; }
     const MD5Digest & Hash() const { return m_Hash; }
     CRecompilerOps *& RecompilerOps() { return m_RecompilerOps; }
@@ -47,19 +47,19 @@ private:
 
     bool AnalyseBlock();
 
-    bool CreateBlockLinkage ( CCodeSection * EnterSection );
-    void DetermineLoops     ();
-    void LogSectionInfo     ();
-    bool SetSection         ( CCodeSection * & Section, CCodeSection * CurrentSection, uint32_t TargetPC, bool LinkAllowed, uint32_t CurrentPC );
-    bool AnalyzeInstruction ( uint32_t PC, uint32_t & TargetPC, uint32_t & ContinuePC, bool & LikelyBranch, bool & IncludeDelaySlot,
-        bool & EndBlock, bool & PermLoop );
+    bool CreateBlockLinkage(CCodeSection * EnterSection);
+    void DetermineLoops();
+    void LogSectionInfo();
+    bool SetSection(CCodeSection * & Section, CCodeSection * CurrentSection, uint32_t TargetPC, bool LinkAllowed, uint32_t CurrentPC);
+    bool AnalyzeInstruction(uint32_t PC, uint32_t & TargetPC, uint32_t & ContinuePC, bool & LikelyBranch, bool & IncludeDelaySlot,
+        bool & EndBlock, bool & PermLoop);
 
     uint32_t           m_VAddrEnter;
     uint32_t           m_VAddrFirst;       // the address of the first opcode in the block
     uint32_t           m_VAddrLast;        // the address of the first opcode in the block
     uint8_t*           m_CompiledLocation; // What address is this compiled at
 
-    typedef std::map<uint32_t,CCodeSection *> SectionMap;
+    typedef std::map<uint32_t, CCodeSection *> SectionMap;
     typedef std::list<CCodeSection *>      SectionList;
 
     SectionMap       m_SectionMap;
